refactor(web): extract shared lane-track drag-and-drop hook
